feat(import): first pass at Nighbot playlist
This commit implements the first pass at importing the playlist managed
by Nightbot. In the code, I have left a few comments to make note of
some undercooked solutions.

I tried doing this with simple `React.useState()` but you may find that
this approach is too naive. For now, I do it this way to make it easier
to output the playlist resource into a text field. It's something like
a `console.log(playlist)` with extra steps.