feat(import): implement first playlist import
We now have a complete cycle starting from authorizing the user with
nightbot, and ending with songs imported directly into an empty sogebot
playlist.

This is still a draft. There is no error handling, but it's very easy
to receive a 429 from the Nightbot API. I didn't see any docs that
described any rate limits, but all errors should at least contain a
`message` key describing the issue. I need to investigate how to
handle errors in a way that would be idiomatically aligned with
the SogeBot source code.