feat(import): less faulty error handling
The previous attempt at error handling was a farce. It did not work at
all because I didn't actually catch the errors. I tested for results,
but those results don't return when the response is a 429 error.

This version tries to use an axios interceptor to do something when
a non-success code is returned. Right now it just prints an error
object and returns undefined (or is it null?).

Some additional effort was made to declare better types for the api
response.