feat(import): Import YouTube playlist from Nightbot (#22)
* feat(import): first pass at Nighbot playlist

This commit implements the first pass at importing the playlist managed
by Nightbot. In the code, I have left a few comments to make note of
some undercooked solutions.

I tried doing this with simple `React.useState()` but you may find that
this approach is too naive. For now, I do it this way to make it easier
to output the playlist resource into a text field. It's something like
a `console.log(playlist)` with extra steps.

* feat(import): improve output and correct comments

* feat(import): implement first playlist import

We now have a complete cycle starting from authorizing the user with
nightbot, and ending with songs imported directly into an empty sogebot
playlist.

This is still a draft. There is no error handling, but it's very easy
to receive a 429 from the Nightbot API. I didn't see any docs that
described any rate limits, but all errors should at least contain a
`message` key describing the issue. I need to investigate how to
handle errors in a way that would be idiomatically aligned with
the SogeBot source code.

* feat(import): Implement basic error handling

This commit additionally makes an effort at fixing the type errors.
One may consider this effort more of an appeasement than solution.

* feat(import): less faulty error handling

The previous attempt at error handling was a farce. It did not work at
all because I didn't actually catch the errors. I tested for results,
but those results don't return when the response is a 429 error.

This version tries to use an axios interceptor to do something when
a non-success code is returned. Right now it just prints an error
object and returns undefined (or is it null?).

Some additional effort was made to declare better types for the api
response.

* feat(import): implement clearer fetching

`fetchResource()` is still recursive, but pagination is heavily
simplified by incrementing the `offset` variable as an argument
to the function.

This also re-enables full playlist importing to make it possible to
encounter errors that I was unable to witness with a small playlist.

I am not confident that an axios interceptor can be applied in order
to handle errors. It seems promising, so maybe I am just doing it
wrong. However, I am ready to dispose of it, because this technique
also forces our usage of `axios.get()` to look unclear while not
providing any actual error handling. These really seem more like
transformers, or maybe a way to verify the shape of a request or
response before committing to processing it.

* feat(import): attempt to retry on 429

* feat(import): remove excessive note comments

* feat(import): remove unused state, fix lint errors

This also messed with the component layout a bit. Since we don't keep
the playlist in state anymore, we can start cleaning up the UI.

* feat(import): improve import failure feedback

* feat(import): replace interceptor with try/catch

* feat(import): remove wrapper fn

* feat(import): Apply suggestions from code review

Co-authored-by: sogehige <sogehige@gmail.com>

* feat(import): return acc from error branches

* feat(import): reimplement retry logic

Previously, the retry logic was naive in a way that made it do nothing.
While it technically did retry a failed request, it only retried a
single time, and the result of the retry was discarded.

This version actually captures error responses and retries each failed
request up to 3 times, delayed by 1 minute each. This will block the
importer from receiving the playlist tracks from Nightbot for
potentially a long time, but will also gracefully handle 429 errors.

Using a generator function may allow for processing a page of tracks
while other requests are awaiting a response. It is unlikely such an
optimization will be necessary, since this feature is expected to be
used only 1 time per user. Additionally, the number of users who will
care to import their playlist from Nightbot is probably very small.

* feat(import): use consistent func definition style

* feat(import): swallow 410 error

* feat(import): output failure as error, not log

---------

Co-authored-by: sogehige <sogehige@gmail.com>