feat(import): convert nb commands to sogebot
The mapping from the properties of nightbot's commands to sogebot's
properties isn't perfect yet.

I had a different problem trying to delete the commands from
`/commands/customcommands/`. All returned an error, but all of the
imported commands were removed from the UI. IIRC it was a 404, which
I can currently recreate by manually adding any new command and
removing it.