fix(import): don't add any nightbot prefixes
The command group will probably suffice as a way to differentiate
these as external commands. No need to make it explicit thrice per
command.

I need to understand how to map nightbot's userLevel property to
sogebot's permission property on the responses. Also, `order` and
`stopIfExecuted` are currently vague to me.