fix(import): generate unique ids on import
This is necessary because I could not find what technique Nightbot
uses to generate its command IDs, but even if I did, I think there
is still a slim chance that some ID could clash with an existing
sogebot command ID. I don't understand cryptography enough to
confirm this, but I think the clash is possible because Nightbot
may be using a weaker (older?) algorithm.

Also, Nightbot either doesn't generate a UUID with dashes in it,
or strips them out. Am I supposed to take them out or leave them in?